change api folder structure by domain
